Stricter typing. Catch2 is now compiled separately from this (externally). Github workflows will need to be modified to reflect this change (allows stricter warnings/errors in compiler). Now safe against -Wconversion and -Wsign-conversion. Need to test on additional hardware, need to update documentation.

+++ include/sac-format/util.hpp
@@ -213,10 +213,10 @@ Trace gen_fake_trace() noexcept {
   sac.kinst("Not Real");
   if (sac.npts() > 0) {
     std::vector<double> data{};
-    data.resize(sac.npts(), 0.0);
+    data.resize(static_cast<size_t>(sac.npts()), 0.0);
     sac.data1(data);
     if (!sac.leven() || sac.iftype() > 1) {
-      data.resize(sac.npts(), 0.0);
+      data.resize(static_cast<size_t>(sac.npts()), 0.0);
       sac.data2(data);
     }
   }
